telco_churn/pyproject.toml
[tool.poetry]
name = "telco-churn"
version = "0.1.0"
description = ""
authors = ["MohammadZeineddine <99921709+MohammadZeineddine@users.noreply.github.com>", "marso16 <marckey2345@gmail.com>"]
readme = "README.md"
packages = [{include = "telco_churn", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<=3.12"
pandas = "^2.2.3"
numpy = "^2.1.3"
scikit-learn = "^1.5.2"
omegaconf = "^2.3.0"
loguru = "^0.7.2"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
mlflow = "^2.18.0"
pydantic = "^2.10.4"
python-multipart = "^0.0.20"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
telco-churn = "telco_churn.core:main"
train-telco = "telco_churn.scripts.train_batch:main"
inference-batch = "telco_churn.scripts.inference_batch:main"
inference-pipeline = "telco_churn.scripts.inference_pipeline_run:main"


[tool.poetry.group.dev.dependencies]
pdoc = "^15.0.0"
mypy = "^1.13.0"
ruff = "^0.7.4"
pytest = "^8.3.3"
invoke = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
pretty = true
strict = true
python_version = ">=3.10,<=3.12"
check_untyped_defs = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.ruff]
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["E", "W", "F", "I"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
